スクリプトについて微調整。
