+++ Atelier_Seed/Assets/Scenes/Miyamoto/CJellyBound.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: af8ae5d6210aebf48983cb8edb89bac8
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
