+++ Atelier_Seed/Library/StateCache/SceneView/5a/5aa8e3155ed3b323e0ad71722b7d68a1.json
@@ -1,1 +1,1 @@
-﻿{"cameraMode":{"drawMode":0,"name":"Shaded","section":"Shading Mode"},"sceneLighting":true,"audioPlay":false,"sceneViewState":{"showFog":true,"showMaterialUpdate":false,"showSkybox":true,"showFlares":true,"showImageEffects":true,"showParticleSystems":true,"showVisualEffectGraphs":true,"m_FxEnabled":true},"in2DMode":true,"pivot":{"x":-9.194999694824219,"y":-2.191999912261963,"z":0.45499998331069949},"rotation":{"x":0.0,"y":0.0,"z":0.0,"w":1.0},"size":9.504695892333985,"orthographic":true}
+﻿{"cameraMode":{"drawMode":0,"name":"Shaded","section":"Shading Mode"},"sceneLighting":true,"audioPlay":false,"sceneViewState":{"showFog":true,"showMaterialUpdate":false,"showSkybox":true,"showFlares":true,"showImageEffects":true,"showParticleSystems":true,"showVisualEffectGraphs":true,"m_FxEnabled":true},"in2DMode":true,"pivot":{"x":-15.793065071105957,"y":-3.314358711242676,"z":-0.11097121238708496},"rotation":{"x":0.0,"y":0.0,"z":0.0,"w":1.0},"size":3.430609941482544,"orthographic":true}
